Rename borg_thread in borg_job. Enable DEBUG for CI.
