+++ src/vorta/borg/info_archive.py
@@ -1,4 +1,4 @@
-from .borg_thread import BorgJob
+from .borg_job import BorgJob
 from vorta.models import ArchiveModel, RepoModel
