+++ src/vorta/borg/info_repo.py
@@ -1,4 +1,4 @@
-from .borg_thread import BorgJob, FakeProfile, FakeRepo
+from .borg_job import BorgJob, FakeProfile, FakeRepo
 from vorta.i18n import trans_late
 from vorta.models import RepoModel
