+++ src/vorta/borg/job_scheduler.py
@@ -5,7 +5,7 @@ from enum import Enum
 from PyQt5 import QtCore
 from PyQt5.QtCore import QObject, QRunnable, QThreadPool
 
-DEBUG = False
+DEBUG = True
 
 
 class JobStatus(Enum):
