+++ src/vorta/borg/mount.py
@@ -1,5 +1,5 @@
 import os
-from .borg_thread import BorgJob
+from .borg_job import BorgJob
 from vorta.models import SettingsModel
