+++ src/vorta/borg/umount.py
@@ -1,6 +1,6 @@
 import os.path
 import psutil
-from .borg_thread import BorgJob
+from .borg_job import BorgJob
 from ..i18n import trans_late
