+++ tests/test_borg.py
@@ -7,7 +7,7 @@ from vorta.borg.prune import BorgPruneJob
 def test_borg_prune(qapp, qtbot, mocker, borg_json_output):
     stdout, stderr = borg_json_output('prune')
     popen_result = mocker.MagicMock(stdout=stdout, stderr=stderr, returncode=0)
-    mocker.patch.object(vorta.borg.borg_thread, 'Popen', return_value=popen_result)
+    mocker.patch.object(vorta.borg.borg_job, 'Popen', return_value=popen_result)
 
     params = BorgPruneJob.prepare(vorta.models.BackupProfileModel.select().first())
     thread = BorgPruneJob(params['cmd'], params, qapp)
