+++ tests/test_repo.py
@@ -3,7 +3,7 @@ import uuid
 import pytest
 from PyQt5 import QtCore
 
-import vorta.borg.borg_thread
+import vorta.borg.borg_job
 import vorta.models
 from vorta.keyring.abc import VortaKeyring
 from vorta.models import EventLogModel, RepoModel, ArchiveModel
@@ -92,7 +92,7 @@ def test_repo_add_success(qapp, qtbot, mocker, borg_json_output):
 
     stdout, stderr = borg_json_output('info')
     popen_result = mocker.MagicMock(stdout=stdout, stderr=stderr, returncode=0)
-    mocker.patch.object(vorta.borg.borg_thread, 'Popen', return_value=popen_result)
+    mocker.patch.object(vorta.borg.borg_job, 'Popen', return_value=popen_result)
 
     add_repo_window.run()
     qtbot.waitUntil(lambda: EventLogModel.select().count() == 2, **pytest._wait_defaults)
@@ -133,7 +133,7 @@ def test_create(qapp, borg_json_output, mocker, qtbot):
     main = qapp.main_window
     stdout, stderr = borg_json_output('create')
     popen_result = mocker.MagicMock(stdout=stdout, stderr=stderr, returncode=0)
-    mocker.patch.object(vorta.borg.borg_thread, 'Popen', return_value=popen_result)
+    mocker.patch.object(vorta.borg.borg_job, 'Popen', return_value=popen_result)
 
     qtbot.mouseClick(main.createStartBtn, QtCore.Qt.LeftButton)
     qtbot.waitUntil(lambda: main.progressText.text().startswith('Backup finished.'), **pytest._wait_defaults)
